leetcode/binary-search/minimum-time-to-complete-trips/main_test.go
package minimum_time_to_complete_trips

import "testing"

func Test_minimumTime(t *testing.T) {
	type args struct {
		time       []int
		totalTrips int
	}
	tests := []struct {
		name string
		args args
		want int64
	}{
		{
			name: "3",
			args: args{time: []int{1, 2, 3}, totalTrips: 5},
			want: 3,
		},
		{
			name: "2",
			args: args{time: []int{2}, totalTrips: 1},
			want: 2,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := minimumTime(tt.args.time, tt.args.totalTrips); got != tt.want {
				t.Errorf("minimumTime() = %v, want %v", got, tt.want)
			}
		})
	}
}
